Refactoring: class GeoDataPointmap::AddPointCommand moved to independent files.
